puppet: runtime-v2 compatibility (#81)

+++ tasks/puppet/test.sh
@@ -1,7 +1,13 @@
 #!/usr/bin/env bash
 
 
-# SERVER_URL='https://my-concord.com' DB_URL='https://puppet-db.com:8081' RBAC_URL='https://puppet-rbac.com:4433' PUPPET_TOKEN='api-token' MOM_CERT='/path/to/cert.pem' ./test.sh
+# SERVER_URL='https://my-concord.com' \
+# DB_URL='https://puppet-db.com:8081' \
+# RBAC_URL='https://puppet-rbac.com:4433' \
+# PUPPET_TOKEN='api-token' \
+# MOM_CERT='/path/to/cert.pem' \
+# RUNTIME=v2 \
+# ./test.sh
 # or
 # PARAM_FILE=/path/to/params.properties ./test.sh
 
@@ -40,21 +46,66 @@ fi
 
 
 
+# see if task jar needs build/rebuild
+doBuild=1
 
-# build the task
-mvn clean package -DskipTests
+if [ -f ./target/last_hash.md5 ]; then
+  echo "Checking for ./src modifications"
+  lastHash=$(cat ./target/last_hash.md5)
+  # md5sum all source files, then md5sum the result
+  currentHash=$(find ./src -type f -exec md5sum {} + | md5sum | awk '{print $1}')
+  if [[ "${lastHash}" == "${currentHash}" ]]; then
+    echo "No src files changed. Skipping build..."
+    doBuild=
+  else
+    echo "${currentHash}" > ./target/last_hash.md5
+  fi
+fi
 
+# build task jar if src files changed
+if [ $doBuild ]; then
+  echo "Building task jar..."
+  # build the task
+  # cd to parent and build to keep from adding apache license
+  pushd ../..
+  mvn clean -pl tasks/puppet package -DskipTests=true
+  popd || exit
+  # save md5 of src files to check next time
+  find ./src -type f -exec md5sum {} + | md5sum | awk '{print $1}' > target/last_hash.md5
+fi
+
+# remove old payload files
 rm -rf target/test && mkdir -p target/test
 # copy the test flow
-cp test.yml target/test/concord.yml
+if [ "${RUNTIME:-}" == "v2" ]; then
+  echo "Using runtime-v2"
+  cp testV2.yml target/test/concord.yml
+else
+  echo "Using runtime-v1"
+  cp test.yml target/test/concord.yml
+fi
 
 # copy the task's JAR...
 mkdir -p target/test/lib/
 cp target/puppet-task-*.jar target/test/lib/
 
 # ...and dependencies
-mvn dependency:copy-dependencies -DoutputDirectory=$(pwd)/target/lib_tmp/ -Dmdep.useSubDirectoryPerScope > /dev/null
+mvn dependency:copy-dependencies -DoutputDirectory="${PWD}/target/lib_tmp/" -Dmdep.useSubDirectoryPerScope > /dev/null
 cp target/lib_tmp/compile/*.jar target/test/lib/
+
+# enable JVM debugging if parameter given and target server is localhost
+if [ "${DEBUG:-}" == "true" ]; then
+  if [[ "${SERVER_URL}" =~ .*"localhost".*  ]]; then
+    echo "Debug enabled"
+    cat > target/test/_agent.json << EOF
+{
+    "jvmArgs": ["-Xdebug", "-Xrunjdwp:transport=dt_socket,address=5005,server=y,suspend=y"]
+}
+EOF
+  else
+    echo "Debug not allowed on non-localhost servers"
+  fi
+fi
 
 
 # copy MoM cert or create an empty one
@@ -64,15 +115,15 @@ else
     cp "${MOM_CERT}" target/test/puppet-mom.pem
 fi
 
-pushd target/test > /dev/null && zip -r payload.zip ./* > /dev/null && popd > /dev/null
+pushd target/test > /dev/null && zip -r payload.zip ./* > /dev/null && popd > /dev/null || exit
 
 HTTP_CODE=$(curl -sn \
   -o target/test/out.json \
   -w '%{http_code}' \
   -F org=Default \
-  -F arguments.puppetParams.databaseUrl=${DB_URL} \
-  -F arguments.puppetParams.rbacUrl=${RBAC_URL} \
-  -F arguments.puppetParams.apiToken=${PUPPET_TOKEN} \
+  -F arguments.puppetParams.databaseUrl="${DB_URL}" \
+  -F arguments.puppetParams.rbacUrl="${RBAC_URL}" \
+  -F arguments.puppetParams.apiToken="${PUPPET_TOKEN}" \
   -F arguments.puppetParams.certificate.path='puppet-mom.pem' \
   -F archive=@target/test/payload.zip \
   "${SERVER_URL}/api/v1/process"
